+++ src/main/java/ru/jamsys/rate/limit/item/RateLimitItemTps.java
@@ -21,7 +21,7 @@ public class RateLimitItemTps implements RateLimitItem {
         return max.get() < 0 || (max.get() > 0 && tps.get() <= max.get()); // -1 = infinity; 0 = reject
     }
 
-    public void setMax(Integer limit) {
+    public void setMax(Long limit) {
         this.max.set(limit);
     }
